Added Login Register Screens and organized accesses to urls

+++ post/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, get_object_or_404,HttpResponseRedirect,redirect
+from django.shortcuts import render, get_object_or_404,HttpResponseRedirect,redirect,Http404
 from .models import Post
 from .forms import PostForm
 from django.contrib import messages
@@ -15,9 +15,15 @@ def post_detail(request,id):
 
 
 def post_create(request):
+    if not request.user.is_authenticated:
+        raise Http404
+
     form = PostForm(request.POST or None,request.FILES or None)
     if form.is_valid():
-       post = form.save()
+       post = form.save(commit=False)
+       post.user = request.user
+       post.save()
+
        messages.success(request,'Post başarılı şekilde oluşturuldu')
        return HttpResponseRedirect(post.get_absolute_url())
 
@@ -27,6 +33,8 @@ def post_create(request):
     return render(request, 'post/form.html', context)
 
 def post_update(request,id):
+    if not request.user.is_authenticated:
+        raise Http404
     post = get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, request.FILES or None,instance=post)
     if form.is_valid():
@@ -40,6 +48,8 @@ def post_update(request,id):
     return render(request, 'post/form.html', context)
 
 def post_delete(request,id):
+    if not request.user.is_authenticated:
+        raise Http404
     post = get_object_or_404(Post, id=id)
     post.delete()
     messages.success(request, 'Post başarılı şekilde silindi',extra_tags='mesaj-basarili')
